Post-rebase fixes
